started on histogram ; moved functions to utils/visualization.ts
